requirements.txt
httpx>=0.25.0
rich>=13.0.0
Pillow>=10.0.0
dnspython>=2.4.0
pydantic>=2.0.0
python-whois>=0.9.0
pytest>=7.0.0
pytest-asyncio>=0.23.0
